+++ src/app/components/elements/ButtonSecondary.tsx
@@ -1,0 +1,8 @@
+import { FC, ReactElement } from "react";
+import "app/styles/component/elements/ButtonSecondary.scss";
+
+const ButtonSecondary: FC = (): ReactElement => {
+  return <button className="button-secondary">Subscribe</button>;
+};
+
+export default ButtonSecondary;
