--- src/app/components/elements/Logo/Logo.scss
@@ -1,13 +1,0 @@
-@import "src/app/styles//GlobalStyles//GlobalStyles.scss";
-
-.logo {
-  display: flex;
-  align-items: center;
-  gap: 14px;
-  p {
-    font-weight: 700;
-    font-size: 30px;
-    line-height: calc(35 / 30);
-    color: $primary-color;
-  }
-}
