--- src/app/components/modules/Bottom/Bottom.tsx
@@ -1,7 +1,0 @@
-import { FC, ReactElement } from "react";
-
-const Bottom: FC = (): ReactElement => {
-  return <h1>Bottom</h1>;
-};
-
-export default Bottom;
