--- src/app/components/modules/Footer/Footer.scss
@@ -1,15 +1,0 @@
-@import "src/app/styles//GlobalStyles/GlobalStyles.scss";
-
-.footer {
-  background-color: $primary-color;
-  height: 70px;
-  padding-top: 26px;
-  padding-bottom: 25px;
-  .container {
-    .footer-content {
-      font-size: 16px;
-      line-height: calc(19 / 16);
-      color: $white-color;
-    }
-  }
-}
