+++ src/app/components/modules/Hero.tsx
@@ -1,0 +1,13 @@
+import { FC, ReactElement } from "react";
+import images from "public/images";
+import "app/styles/component/modules/Hero.scss";
+
+const Hero: FC = (): ReactElement => {
+  return (
+    <section className="hero">
+      <img src={images.hero} alt="" />
+    </section>
+  );
+};
+
+export default Hero;
