--- src/app/components/modules/Hero/Hero.scss
@@ -1,1 +1,0 @@
-@import "src/app/styles//GlobalStyles//GlobalStyles.scss";
