+++ src/app/const/reasons.ts
@@ -1,0 +1,23 @@
+import { ReasonBoxProps } from "app/components/elements/ReasonBox";
+import images from "public/images";
+
+export const listReason: ReasonBoxProps[] = [
+  {
+    id: 1,
+    imageUrl: images.reasonIcon1,
+    title: "Best Services",
+    desc: "Nullam senectus porttitor in eget. Eget rutrum leo interdum.",
+  },
+  {
+    id: 2,
+    imageUrl: images.reasonIcon1,
+    title: "Best Teams",
+    desc: "Cursus semper tellus volutpat aliquet lacus.",
+  },
+  {
+    id: 3,
+    imageUrl: images.reasonIcon2,
+    title: "Best Designs",
+    desc: "Ultricies at ipsum nunc, tristique nam lectus.",
+  },
+];
