+++ src/app/const/services.ts
@@ -1,0 +1,41 @@
+import { ServiceCardProps } from "app/components/elements/ServiceCard";
+import images from "public/images";
+
+export const listService: ServiceCardProps[] = [
+  {
+    id: 1,
+    imageUrl: images.serviceIcon1,
+    service: "Construction",
+    theme: "white",
+  },
+  {
+    id: 2,
+    imageUrl: images.serviceIcon2,
+    service: "Renovation",
+    theme: "blue",
+  },
+  {
+    id: 3,
+    imageUrl: images.serviceIcon3,
+    service: "Construction",
+    theme: "white",
+  },
+  {
+    id: 4,
+    imageUrl: images.serviceIcon4,
+    service: "Repair Services",
+    theme: "blue",
+  },
+  {
+    id: 5,
+    imageUrl: images.serviceIcon5,
+    service: "Architecture",
+    theme: "white",
+  },
+  {
+    id: 6,
+    imageUrl: images.serviceIcon6,
+    service: "Electric",
+    theme: "blue",
+  },
+];
