+++ src/app/styles/GlobalStyles/GlobalStyles.scss
@@ -3,7 +3,6 @@ $primary-color: #2947a9;
 $white-color: #ffffff;
 $secondary-color: #f9995d;
 
-//
 .container {
   max-width: 1216px;
   margin: 0 auto;
