+++ src/app/styles/component/elements/ButtonSecondary.scss
@@ -1,0 +1,17 @@
+@import "src/app/styles/GlobalStyles/GlobalStyles.scss";
+
+.button-secondary {
+  width: 136px;
+  height: 41px;
+  // padding: 16px 20px;
+  gap: 10px;
+  background: $secondary-color;
+  outline: none;
+  border: none;
+  border-radius: 4px;
+  color: $white-color;
+  font-weight: 600;
+  font-size: 18px;
+  line-height: calc(21 / 18);
+  text-align: center;
+}
