+++ src/app/styles/component/elements/ReasonBox.scss
