+++ src/app/styles/component/elements/ServiceCard.scss
@@ -1,0 +1,38 @@
+@import "src/app/styles/GlobalStyles/GlobalStyles.scss";
+
+.theme-white {
+  background-color: $white-color !important;
+}
+
+.primary-color {
+  color: $primary-color !important;
+}
+
+.card {
+  width: 271px;
+  height: 181px;
+  background-color: #2947a9;
+  box-shadow: 0px 0px 37px -2px rgba(17, 24, 39, 0.05);
+  border-radius: 4px;
+
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 20px;
+  gap: 20px;
+  &-icon {
+    width: 40px;
+  }
+  &-bulkhead {
+    width: 60px;
+    height: 2px;
+    background-color: #e0e3eb;
+  }
+  &-sevice {
+    font-weight: 600;
+    font-size: 20px;
+    line-height: calc(23 / 20);
+    color: #e0e3eb;
+  }
+}
