+++ src/app/styles/component/elements/SocialGroup.scss
@@ -1,0 +1,7 @@
+@import "src/app/styles/GlobalStyles/GlobalStyles.scss";
+
+.social {
+  display: flex;
+  align-items: center;
+  gap: 21px;
+}
