+++ src/app/styles/component/modules/Bottom.scss
@@ -1,0 +1,85 @@
+@import "src/app/styles/GlobalStyles/GlobalStyles.scss";
+
+.bottom {
+  padding-top: 85px;
+  margin-bottom: 80px;
+  .container {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 323px;
+    .bottom-left {
+      .bottom-contact {
+        width: 469px;
+        height: 133px;
+        display: flex;
+        flex-direction: column;
+        align-items: flex-start;
+        gap: 32px;
+        margin-bottom: 32px;
+        .item-title {
+          font-weight: 500;
+          font-size: 20px;
+          line-height: calc(23 / 20);
+          color: $primary-color;
+          text-transform: uppercase;
+          font-variant: small-caps;
+        }
+        .item-content {
+          font-size: 20px;
+          line-height: calc(23 / 20);
+          text-transform: capitalize;
+        }
+        .contact-address {
+          display: flex;
+          flex-direction: row;
+          align-items: flex-start;
+          padding: 0px;
+          gap: 13px;
+        }
+        .contact-phone {
+          display: flex;
+          flex-direction: row;
+          align-items: flex-start;
+          padding: 0px;
+          gap: 34px;
+        }
+        .contact-email {
+          display: flex;
+          flex-direction: row;
+          align-items: flex-start;
+          padding: 0px;
+          gap: 41px;
+        }
+      }
+    }
+    .bottom-right {
+      .bottom-news {
+        margin-bottom: 42px;
+        .news-title {
+          font-weight: 500;
+          font-size: 20px;
+          line-height: calc(23 / 20);
+          font-variant: small-caps;
+          color: $primary-color;
+          margin-bottom: 20px;
+        }
+        .new-content {
+          display: flex;
+          align-items: center;
+          gap: 8px;
+        }
+      }
+      .bottom-social {
+        font-weight: 500;
+        font-size: 20px;
+        line-height: calc(23 / 20);
+        font-variant: small-caps;
+        color: $primary-color;
+        text-transform: uppercase;
+        .social-title {
+          margin-bottom: 20px;
+        }
+      }
+    }
+  }
+}
