+++ src/app/styles/component/modules/Footer.scss
@@ -1,0 +1,15 @@
+@import "src/app/styles/GlobalStyles/GlobalStyles.scss";
+
+.footer {
+  background-color: $primary-color;
+  height: 70px;
+  padding-top: 26px;
+  padding-bottom: 25px;
+  .container {
+    .footer-content {
+      font-size: 16px;
+      line-height: calc(19 / 16);
+      color: $white-color;
+    }
+  }
+}
