+++ src/app/styles/component/modules/Hero.scss
@@ -1,0 +1,1 @@
+@import "src/app/styles/GlobalStyles/GlobalStyles.scss";
