+++ src/app/styles/component/modules/Navbar.scss
@@ -1,0 +1,14 @@
+@import "src/app/styles/GlobalStyles/GlobalStyles.scss";
+
+.active {
+  color: $secondary-color !important;
+  font-weight: 600 !important;
+}
+.navbar {
+  .container {
+    height: 80px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+  }
+}
