+++ src/app/styles/component/modules/Reasons.scss
@@ -1,0 +1,22 @@
+@import "src/app/styles/GlobalStyles/GlobalStyles.scss";
+
+.reasons {
+  .container {
+    .reasons-title {
+      font-weight: 700;
+      font-size: 36px;
+      line-height: calc(42 / 36);
+      color: #292e3d;
+      text-align: center;
+      margin-top: 60px;
+      margin-bottom: 50px;
+    }
+    .reasons-content {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      gap: 72px;
+      margin-bottom: 84px;
+    }
+  }
+}
