--- src/app/types/Type.tsx
@@ -1,13 +1,0 @@
-export type ReasonType = {
-  id: number;
-  imageUrl: string;
-  title: string;
-  desc: string;
-};
-
-export type ServiceType = {
-  id: number;
-  imageUrl: string;
-  service: string;
-  theme: string;
-};
